Distributed/user_service/requirements.txt
asgiref
atlastk
cffi
click
cryptography
Cython
ecdsa
fastapi
FastAPI-SQLAlchemy
greenlet
h11
joblib
pyasn1
pycparser
pydantic
python-jose
python-multipart
rsa
six
SQLAlchemy
starlette
threadpoolctl
tortoise
typing-extensions
uvicorn
passlib
bcrypt
psycopg2-binary
python-dotenv